Change /getlists route to return array of lists
previosly /getlists generated an object with an array of lists as well as count of notes in each list. This was time consuming and slowed down the performance of /getlists response, also affecting the performance of the App front-end waiting on the lists.

getcount to be possibly added in future as seperate API call.